soulCare/soulcare_backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2

# --- New Chat & ASGI Dependencies ---
channels==4.0.0
channels_redis==4.2.0
daphne==4.1.2
cryptography==43.0.0